[Modify] 本文のボリューム減量のためopencv_rosのコードを1つにまとめ、起動用launchファイルを追加。
