+++ chapter7/test_opencv/rviz/opencv_ros.rviz
@@ -1,0 +1,116 @@
+Panels:
+  - Class: rviz/Displays
+    Help Height: 75
+    Name: Displays
+    Property Tree Widget:
+      Expanded: ~
+      Splitter Ratio: 0.551501989
+    Tree Height: 362
+  - Class: rviz/Selection
+    Name: Selection
+  - Class: rviz/Tool Properties
+    Expanded:
+      - /2D Pose Estimate1
+      - /2D Nav Goal1
+      - /Publish Point1
+    Name: Tool Properties
+    Splitter Ratio: 0.588679016
+  - Class: rviz/Views
+    Expanded:
+      - /Current View1
+    Name: Views
+    Splitter Ratio: 0.5
+  - Class: rviz/Time
+    Experimental: false
+    Name: Time
+    SyncMode: 0
+    SyncSource: Image_raw
+Visualization Manager:
+  Class: ""
+  Displays:
+    - Class: rviz/Image
+      Enabled: true
+      Image Topic: /camera/image_raw
+      Max Value: 1
+      Median window: 5
+      Min Value: 0
+      Name: Image_raw
+      Normalize Range: true
+      Queue Size: 2
+      Transport Hint: raw
+      Unreliable: false
+      Value: true
+    - Class: rviz/Image
+      Enabled: true
+      Image Topic: /image_converter/output_video
+      Max Value: 1
+      Median window: 5
+      Min Value: 0
+      Name: Image_drawn
+      Normalize Range: true
+      Queue Size: 2
+      Transport Hint: raw
+      Unreliable: false
+      Value: true
+  Enabled: true
+  Global Options:
+    Background Color: 48; 48; 48
+    Default Light: false
+    Fixed Frame: ""
+    Frame Rate: 30
+  Name: root
+  Tools:
+    - Class: rviz/Interact
+      Hide Inactive Objects: true
+    - Class: rviz/MoveCamera
+    - Class: rviz/Select
+    - Class: rviz/FocusCamera
+    - Class: rviz/Measure
+    - Class: rviz/SetInitialPose
+      Topic: /initialpose
+    - Class: rviz/SetGoal
+      Topic: /move_base_simple/goal
+    - Class: rviz/PublishPoint
+      Single click: true
+      Topic: /clicked_point
+  Value: true
+  Views:
+    Current:
+      Angle: 0
+      Class: rviz/TopDownOrtho
+      Enable Stereo Rendering:
+        Stereo Eye Separation: 0.0599999987
+        Stereo Focal Distance: 1
+        Swap Stereo Eyes: false
+        Value: false
+      Invert Z Axis: false
+      Name: Current View
+      Near Clip Distance: 0.00999999978
+      Scale: 10
+      Target Frame: <Fixed Frame>
+      Value: TopDownOrtho (rviz)
+      X: -0.0992676243
+      Y: -0.296530455
+    Saved: ~
+Window Geometry:
+  Displays:
+    collapsed: true
+  Height: 489
+  Hide Left Dock: false
+  Hide Right Dock: true
+  Image_drawn:
+    collapsed: false
+  Image_raw:
+    collapsed: false
+  QMainWindow State: 000000ff00000000fd0000000400000000000000fe000001a3fc0200000009fb000000120049006d006100670065005f0072006100770100000028000000cf0000001600fffffffb000000160049006d006100670065005f0064007200610077006e01000000fd000000ce0000001600fffffffb0000001200530065006c0065006300740069006f006e00000001e10000009b0000006100fffffffb0000001e0054006f006f006c002000500072006f007000650072007400690065007302000001ed000001df00000185000000a3fb000000120056006900650077007300200054006f006f02000001df000002110000018500000122fb000000200054006f006f006c002000500072006f0070006500720074006900650073003203000002880000011d000002210000017afb0000002000730065006c0065006300740069006f006e00200062007500660066006500720200000138000000aa0000023a00000294fb00000014005700690064006500530074006500720065006f02000000e6000000d2000003ee0000030bfb0000000c004b0069006e0065006300740200000186000001060000030c000002610000000100000300000002a2fc0200000005fb0000001e0054006f006f006c002000500072006f00700065007200740069006500730100000041000000780000000000000000fb000000100044006900730070006c0061007900730000000028000000d9000000d700fffffffb0000000a00560069006500770073000000010700000170000000ad00fffffffb0000000800540069006d0065000000027d0000004d0000003b00fffffffb0000001200530065006c0065006300740069006f006e010000025a000000b200000000000000000000000200000515000000d7fc0100000001fb0000000a00560069006500770073030000004e00000080000002e1000001970000000300000515000000d7fc0100000001fb0000000800540069006d0065010000000000000450000000000000000000000021000001a300000004000000040000000800000008fc0000000100000002000000010000000a0054006f006f006c00730100000000ffffffff0000000000000000
+  Selection:
+    collapsed: false
+  Time:
+    collapsed: true
+  Tool Properties:
+    collapsed: false
+  Views:
+    collapsed: true
+  Width: 293
+  X: 708
+  Y: 30
